Introduced GenericCRUDElement and deprecated AbstractCRUD for consistency with other elements
