+++ spec/FSi/Bundle/AdminBundle/Factory/Worker/FormWorkerSpec.php
@@ -3,6 +3,7 @@
 namespace spec\FSi\Bundle\AdminBundle\Factory\Worker;
 
 use FSi\Bundle\AdminBundle\Admin\CRUD\AbstractCRUD;
+use FSi\Bundle\AdminBundle\Admin\CRUD\GenericCRUDElement;
 use PhpSpec\ObjectBehavior;
 use Symfony\Component\Form\FormFactoryInterface;
 
@@ -14,7 +15,7 @@ class FormWorkerSpec extends ObjectBehavior
     }
 
     function it_mount_form_factory_to_elements_that_are_form_aware(
-        AbstractCRUD $element,
+        GenericCRUDElement $element,
         FormFactoryInterface $formFactory
     ) {
         $element->setFormFactory($formFactory)->shouldBeCalled();
